removing references to JPL
I haven't yet actually tried this out on a non-JPL server. The way it's set up assumes you're using JPL's institutional Apache hosting--it should work in a normal Apache hosting environment as long as your webroot is in `www` and you have a folder named `private` which is parallel to it.